requirements.txt
# Core Django
asgiref==3.11.0
Django==5.2.10
django-environ==0.12.0

# Database
psycopg2-binary==2.9.11

# Frontend & Forms
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
pillow==12.1.0

# Email & External Services
resend==2.30.0

# Utilities
django-mathfilters==1.0.0
reportlab==4.4.9
sqlparse==0.5.5
charset-normalizer==3.4.4
psutil==6.0.0
tzdata==2025.3

# Production Server
gunicorn==23.0.0

# Static Files & Security
whitenoise==6.11.0

# Optional: For API Documentation (if needed)
# djangorestframework==3.14.0
# drf-spectacular==0.27.0

# Optional: For Caching
# redis==5.0.1

# Optional: For Task Queue
# celery==5.4.0

# Cloudflare R2 storage dependencies
boto3==1.30.0
django-storages[boto3]==1.19.1
dj-database-url==2.1.0
